+++ function.h
@@ -1,5 +1,5 @@
-int addition(int S,int l);
-int subtraction(int S,int l);
-int multiplication(int S,int l);
-int division(int S,int l);
+double addition(double S,double l);
+double subtraction(double S,double l);
+double multiplication(double S,double l);
+double division(double S,double l);
